+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Configurable item sales, whitelisting, and blacklisting.
 - Minor fixes for `Terminal.RotateShipDecorSelection()` patch.
 	- `maxItems` value is now used instead of `minItems` when `minItems` is greater than `maxItems`, as intended.
 	- Fixed items not sorting alphabetically with `stockAll` disabled.
+- Everything should still be compatible with every game version since `v45`.
 - Updated `LICENSE` name and year.
 
 ### [2.2.1]